pyproject.toml
[project]
name = "orange-hrm-robotframework"
version = "1.0.0"
description = "A Robot Framework project for automating Orange HRM application testing."
authors = [
    {name = "Mateo Agudelo Garcia",email = "mateo.agudelo80@eia.edu.co"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "robotframework-seleniumlibrary (>=6.7.1,<7.0.0)",
    "robotframework-tidy (>=4.16.0,<5.0.0)",
    "robotframework-retryfailed (>=0.2.0,<0.3.0)",
    "faker (>=37.0.2,<38.0.0)",
]

[tool.poetry]
packages = [
    {include = "PageObject"}
]

[tool.poetry.group.dev.dependencies]
robotframework = "^7.2.2"
pre-commit = "^4.1.0"
pylint = "^3.3.5"
flake8 = "^7.1.2"
robotframework-robocop = "^5.8.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
robot-tests = "run_robot_tests:main"
